+++ app/(entrance)/login/components/LoginPage.js
@@ -1,6 +1,6 @@
 'use client'
 import EmailInput, { isValidEmail } from '@shared/inputs/EmailInput.js'
-import PasswordInput from '@shared/inputs/PasswrodInput.js'
+import PasswordInput from '@shared/inputs/PasswordInput.js'
 import { setUserCookie } from 'helpers/cookieHelper.js'
 import Link from 'next/link.js'
 import { Suspense, useState } from 'react'
