+++ app/(entrance)/sign-up/components/SignUpPage.js
@@ -1,7 +1,7 @@
 'use client'
 import { isValidEmail } from '@shared/inputs/EmailInput.js'
 import { isValidPhone } from '@shared/inputs/PhoneInput.js'
-import PasswordInput from '@shared/inputs/PasswrodInput.js'
+import PasswordInput from '@shared/inputs/PasswordInput.js'
 import MaxWidth from '@shared/layouts/MaxWidth'
 import { Fragment, useState } from 'react'
 import H1 from '@shared/typography/H1'
