+++ app/(user)/profile/components/PasswordSection.js
@@ -6,7 +6,7 @@ import Body2 from '@shared/typography/Body2'
 import PrimaryButton from '@shared/buttons/PrimaryButton'
 import Paper from '@shared/utils/Paper'
 import H2 from '@shared/typography/H2'
-import PasswordInput from '@shared/inputs/PasswrodInput'
+import PasswordInput from '@shared/inputs/PasswordInput'
 import DeleteAccountButton from './DeleteAccountButton'
 import { apiRoutes } from 'gpApi/routes'
 import { clientFetch } from 'gpApi/clientFetch'
